refactor: 💡 move code to communicate with node to client.rs
